Move post-save hooks to ContentsManager, add deprecation warnings, add hook override warnings
